+++ demisto_sdk/commands/content_graph/objects/__init__.py
@@ -33,7 +33,11 @@ __all__ = [
     "RelationshipData",
     "LayoutRule",
     "PreProcessRule",
+    "AssetsModelingRule",
 ]
+from demisto_sdk.commands.content_graph.objects.assets_modeling_rule import (
+    AssetsModelingRule,
+)
 from demisto_sdk.commands.content_graph.objects.base_playbook import BasePlaybook
 from demisto_sdk.commands.content_graph.objects.classifier import Classifier
 from demisto_sdk.commands.content_graph.objects.correlation_rule import CorrelationRule