+++ demisto_sdk/commands/coverage_analyze/coverage_report.py
@@ -105,6 +105,7 @@ class CoverageReport:
             logger.debug(
                 f"coverage report {self.cov.config.data_file} file not found. "
             )
+            raise FileNotFoundError("The coverage file does not exist.")
 
         logger.info(f"\n{self.report_str}")
 
@@ -134,8 +135,8 @@ class CoverageReport:
             min_cov = self.file_min_coverage(file_name)
             if min_cov > cov_precents:
                 logger.error(
-                    f"[red]Unit-tests for '{file_name}' must reach a coverage of at least {min_cov}% "
-                    f"(currently: {cov_precents}%).[/red]"
+                    f"[red]Unit-tests for '{file_name}' must reach a coverage of at least {min_cov} percent "
+                    f"(currently: {cov_precents} percent).[/red]"
                 )
                 coverage_ok = False
         return coverage_ok