+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 ## Unreleased
+* Fixed an issue in **prepare-content** where tags were added to metadata because of test scripts.
 
 ## 1.23.0
 * Added support for inputs sections and outputs sections in a playbook.