+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 ## Unreleased
+* Fixed an issue where the error was not clear when trying to retrieve the server version.
+* Fixed an issue in **prepare-content** where tags were added to metadata because of test scripts.
+* fixed an issue where to_id_set_entity method failed on id extraction due to missing pack.
 * Fixed an issue where **run-playbook** command did not work.
 
 ## 1.23.0